+++ src/main/java/me/darki/konas/Class236.java
@@ -116,10 +116,10 @@ extends Class215 {
         block0: {
             block1: {
                 super.Method497(n, n2, f);
-                GL11.glEnable((int)3089);
+                GL11.glEnable(3089);
                 RenderUtil2.Method1339(new ScaledResolution(Minecraft.getMinecraft()), this.Method486(), this.Method492(), this.Method489(), this.Method476());
                 this.Field2506.forEach(arg_0 -> Class236.Method2164(n, n2, f, arg_0));
-                GL11.glDisable((int)3089);
+                GL11.glDisable(3089);
                 if (!this.Method501(n, n2)) break block0;
                 if (Field327 <= 0) break block1;
                 if (!(this.Field2506.get(0).Method495() < 0.0f)) break block0;
