+++ src/main/java/me/darki/konas/Class473.java
@@ -34,13 +34,13 @@ implements EventListener {
             exception.printStackTrace();
             return 20.0f;
         }
-        return MathHelper.clamp((float)f3, (float)f2, (float)f);
+        return MathHelper.clamp(f3, f2, f);
     }
 
     public void Method2191() {
         if (this.Field2560 != -1L) {
             float f = (float)(System.currentTimeMillis() - this.Field2560) / 1000.0f;
-            this.Field2558[this.Field2559 % this.Field2558.length] = MathHelper.clamp((float)(20.0f / f), (float)0.0f, (float)20.0f);
+            this.Field2558[this.Field2559 % this.Field2558.length] = MathHelper.clamp(20.0f / f, 0.0f, 20.0f);
             ++this.Field2559;
         }
         this.Field2560 = System.currentTimeMillis();
@@ -52,7 +52,7 @@ implements EventListener {
             if (!(f2 > 0.0f) || !(f2 < f)) continue;
             f = f2;
         }
-        return MathHelper.clamp((float)f, (float)0.0f, (float)20.0f);
+        return MathHelper.clamp(f, 0.0f, 20.0f);
     }
 
     public float Method2193() {
@@ -63,7 +63,7 @@ implements EventListener {
             f2 += f3;
             f += 1.0f;
         }
-        return MathHelper.clamp((float)(f2 / f), (float)0.0f, (float)20.0f);
+        return MathHelper.clamp(f2 / f, 0.0f, 20.0f);
     }
 
     public Class473() {
